+++ blocks/articles-rail/articles-rail.js
@@ -1,61 +1,6 @@
-import ffetch from '../../scripts/ffetch.js';
 import { getMetadata } from '../../scripts/lib-franklin.js';
-import { getLink, getPathDetails, createPlaceholderSpan, htmlToElement } from '../../scripts/scripts.js';
-
-/**
- * Helper function thats returns a list of all products
- * - below <lang>/browse/<product-page>
- * - To get added, the product page must be published
- * - Product pages listed in <lang>/browse/top-products are put at the the top
- *   in the order they appear in top-products
- * - the top product list can point to sub product pages
- */
-export async function getProducts() {
-  // get language
-  const { lang } = getPathDetails();
-  let featured = true;
-  const [Products, publishedPages] = await Promise.all([
-    // load the <lang>/top-product list
-    ffetch(`/${lang}/top-products.json`, `/en/top-products.json`).all(),
-    // get all indexed pages below <lang>/browse
-    ffetch(`/${lang}/browse-index.json`, `/en/browse-index.json`).all(),
-  ]);
-
-  // add all published top products to final list
-  const finalProducts = Products.filter((product) => {
-    // if separator is reached
-    if (product.path.startsWith('-')) {
-      featured = false;
-      return false;
-    }
-
-    // check if product is in published list
-    const found = publishedPages.find((elem) => elem.path === product.path);
-    if (found) {
-      // keep original title if no nav title is set
-      if (!product.title) product.title = found.title;
-      // set featured flag
-      product.featured = featured;
-      // remove it from publishedProducts list
-      publishedPages.splice(publishedPages.indexOf(found), 1);
-      return true;
-    }
-    return false;
-  });
-
-  // if no separator was found , add the remaining products alphabetically
-  if (featured) {
-    // for the rest only keep main product pages (<lang>/browse/<main-product-page>)
-    const publishedMainProducts = publishedPages
-      .filter((page) => page.path.split('/').length === 4)
-      // sort alphabetically
-      .sort((productA, productB) => productA.path.localeCompare(productB.path));
-    // append remaining published products to final list
-    finalProducts.push(...publishedMainProducts);
-  }
-
-  return finalProducts;
-}
+import { getLink, createPlaceholderSpan, htmlToElement } from '../../scripts/scripts.js';
+import getProducts from '../../scripts/utils/product-utils.js';
 
 // Utility function to toggle visibility of items
 function toggleItemVisibility(itemList, startIndex, show) {
@@ -92,7 +37,7 @@ function handleViewLessClick(block, numFeaturedProducts) {
 }
 
 async function displayAllProducts(block) {
-  const productList = await getProducts();
+  const productList = await getProducts('articles');
 
   if (productList.length > 0) {
     const productsUL = document.createElement('ul');
@@ -169,8 +114,7 @@ async function displayAllProducts(block) {
   }
 }
 
-function displayManualNav(manualNav, block) {
-  // get root ul
+function displayManualNav(block, manualNav) {
   const rootUL = manualNav.querySelector('ul');
   rootUL.classList.add('subPages');
 
@@ -199,20 +143,11 @@ function displayManualNav(manualNav, block) {
 
 // Main function to decorate the block
 export default async function decorate(block) {
-  // get any defined manual navigation
   const [manualNav] = block.querySelectorAll(':scope div > div');
-
-  // to avoid dublication when editing
   block.textContent = '';
 
   const theme = getMetadata('theme');
 
-  // const label = getMetadata('og:title');
-
-  // const results = await ffetch(`/${getPathDetails().lang}/articles-index.json`).all();
-  // const currentPagePath = getEDSLink(window.location.pathname);
-
-  // For Browse All Page
   if (theme.startsWith('article-')) {
     // Browse By
     const browseByUL = document.createElement('ul');
@@ -232,7 +167,7 @@ export default async function decorate(block) {
     block.append(browseByUL);
     // Show All Products
     if (manualNav) {
-      displayManualNav(manualNav, block);
+      displayManualNav(block, manualNav);
     } else {
       displayAllProducts(block);
     }
